requirements.txt
python-telegram-bot>=20.6
asyncio>=3.4.3 